rename landing => ppm landing
