+++ src/scenes/Ppm/MoveSummary/DraftMoveSummary.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ProfileStatusTimeline } from 'scenes/Landing/StatusTimeline';
+import { ProfileStatusTimeline } from 'scenes/Ppm/StatusTimeline';
 import truck from 'shared/icon/truck-gray.svg';
 
 const DraftMoveSummary = (props) => {
