+++ src/scenes/Ppm/MoveSummary/PpmMoveDetails.module.scss
