+++ src/scenes/Ppm/images/ppm-approved.png
