+++ src/scenes/Ppm/index.jsx
@@ -30,7 +30,7 @@ import { selectActiveOrLatestOrders, selectUploadsForActiveOrders } from 'shared
 import { loadMTOShipments, selectMTOShipmentForMTO } from 'shared/Entities/modules/mtoShipments';
 import { selectActiveOrLatestMove } from 'shared/Entities/modules/moves';
 
-export class Landing extends Component {
+export class PpmLanding extends Component {
   componentDidMount() {
     // Load user into entities
     const { isLoggedIn, showLoggedInUser } = this.props;
@@ -174,7 +174,7 @@ export class Landing extends Component {
   }
 }
 
-Landing.propTypes = {
+PpmLanding.propTypes = {
   context: PropTypes.shape({
     flags: PropTypes.shape({
       hhgFlow: PropTypes.bool,
@@ -183,7 +183,7 @@ Landing.propTypes = {
   }).isRequired,
 };
 
-Landing.defaultProps = {
+PpmLanding.defaultProps = {
   context: {
     flags: {
       hhgFlow: false,
@@ -231,4 +231,4 @@ function mapDispatchToProps(dispatch) {
   );
 }
 
-export default withContext(withLastLocation(connect(mapStateToProps, mapDispatchToProps)(Landing)));
+export default withContext(withLastLocation(connect(mapStateToProps, mapDispatchToProps)(PpmLanding)));
